Add displayContent function
