+++ src/contact_us.js
@@ -7,8 +7,8 @@ const reachUs = () => {
   homeTitle.innerText = "Find Us";
   homeNote.innerText = "Tel: +237679260797\n\n Email: fredtimermih@gmail.com";
 
-  contactUs.setAttribute("class", "main");
-  contactUs.setAttribute("id", "contactsPage");
+  contactUs.setAttribute("class", "section");
+  contactUs.setAttribute("id", "myContacts");
   homeTitle.setAttribute("class", "main-title");
   homeNote.setAttribute("class", "main-quote");
   frame.setAttribute("class", "frame");
