+++ src/index.js
@@ -1,61 +1,32 @@
+import navigationBar from "./navigationbar.js"
 import displayHome from "./home.js";
 import displayMenu from "./menus.js";
 import contactUs from "./contact_us.js"
+import displayContent from "./content.js"
 import myFooter from "./footer.js";
 
-const navigationBar = () => {
-  const navBar = document.createElement("nav");
-  const navHeader = document.createElement("h1");
-  const navUlist = document.createElement("ul");
-  const homeLi = document.createElement("li");
-  const menuLi = document.createElement("li");
-  const contactLi = document.createElement("li");
-  const homeLink = document.createElement("a");
-  const menuLink = document.createElement("a");
-  const contactLink = document.createElement("a");
-
-  navHeader.innerText = "A Taste of Africa - Recipes";
-  homeLink.innerText = "Home";
-  menuLink.innerText = "Menu";
-  contactLink.innerText = "Contact";
-
-  navBar.setAttribute("class", "nav-bar");
-  navHeader.setAttribute("class", "rest-header");
-  navUlist.setAttribute("class", "menu-list");
-  homeLi.setAttribute("class", "menu-item");
-  menuLi.setAttribute("class", "menu-item");
-  contactLi.setAttribute("class", "menu-item");
-  homeLink.setAttribute("class", "menu-link");
-  menuLink.setAttribute("class", "menu-link");
-  contactLink.setAttribute("class", "menu-link");
-
-  homeLink.setAttribute("id", "home-link");
-  menuLink.setAttribute("id", "menu-link");
-  contactLink.setAttribute("id", "contact-link");
-
-  homeLink.setAttribute("href", "#");
-  menuLink.setAttribute("href", "#");
-  contactLink.setAttribute("href", "#");
-
-  navBar.appendChild(navHeader);
-  navBar.appendChild(navUlist);
-  navUlist.appendChild(homeLi);
-  navUlist.appendChild(menuLi);
-  navUlist.appendChild(contactLi);
-  homeLi.appendChild(homeLink);
-  menuLi.appendChild(menuLink);
-  contactLi.appendChild(contactLink);
-
-  return navBar;
-};
 
 const run = () => {
   const content = document.getElementById("content");
   content.appendChild(navigationBar());
-  content.appendChild(myFooter());
   content.appendChild(displayHome());
   content.appendChild(displayMenu());
   content.appendChild(contactUs());
+
+  content.appendChild(myFooter());
+  
+
+  document.getElementById("home-link").onclick = () => {
+    displayContent("homePage");
+  };
+  document.getElementById("menu-link").onclick = () => {
+    displayContent("myMenu");
+  };
+  document.getElementById("contact-link").onclick = () => {
+    displayContent("myContacts");
+  };
+ 
+  
 };
 
 run();
